prevent re-rendering guide content when viewing a pull while content loads
